Moved recaptcha files from root folder to addons folder
