+++ src/main/java/com/train4game/social/addons/recaptcha/ReCaptchaResponseFilter.java
@@ -1,4 +1,4 @@
-package com.train4game.social.recaptcha;
+package com.train4game.social.addons.recaptcha;
 
 import org.springframework.stereotype.Component;
