+++ src/main/java/com/train4game/social/service/RecaptchaService.java
@@ -1,8 +1,8 @@
 package com.train4game.social.service;
 
 import com.train4game.social.model.Recaptcha;
-import com.train4game.social.recaptcha.ReCaptchaCacheService;
-import com.train4game.social.recaptcha.ReCaptchaException;
+import com.train4game.social.addons.recaptcha.ReCaptchaCacheService;
+import com.train4game.social.addons.recaptcha.ReCaptchaException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
