+++ src/main/java/com/train4game/social/to/RegisterUserTo.java
@@ -1,6 +1,6 @@
 package com.train4game.social.to;
 
-import com.train4game.social.recaptcha.ValidReCaptcha;
+import com.train4game.social.addons.recaptcha.ValidReCaptcha;
 import com.train4game.social.web.validators.StringFieldsMatch;
 import com.train4game.social.web.validators.UniqueEmail;
 import lombok.*;
